Task-2
Here is my submission of Task-2 of Data Science Internship at The Sparks Foundation. The objective of this task was -"To Explore Supervised Machine Learning". In this task, we had a dataset(not so big) contains information about the Scores a student gets after studying for a specific Hours. I used LinearRegression to develop a simple model which predict the Marks, a student can score after studying given hours.